10.10.2023 dao findByID

+++ app/src/main/java/com/example/constructor/dao/ChapterReaderWriter.java
@@ -6,6 +6,6 @@ public interface ChapterReaderWriter extends ChapterReader{
 
     long insert (Chapter chapter);
 
-    void update (int id, boolean accepted);
+    void update (long id, boolean accepted);
 
 }
