+++ app/src/main/java/com/example/constructor/model/Test.java
@@ -3,13 +3,13 @@ package com.example.constructor.model;
 import java.util.List;
 
 public class Test {
-    private int id;
+    private long id;
     private String name;
-    private int chapter_id;
+    private long chapter_id;
     private List<Question> questionList;
     private boolean isDone;
 
-    public Test(int id, String name, int chapter_id, List<Question> questionList, boolean isDone) {
+    public Test(long id, String name, long chapter_id, List<Question> questionList, boolean isDone) {
         this.id = id;
         this.name = name;
         this.chapter_id = chapter_id;
@@ -17,14 +17,14 @@ public class Test {
         this.isDone = isDone;
     }
 
-    public Test(String name, int chapter_id, List<Question> questionList, boolean isDone) {
+    public Test(String name, long chapter_id, List<Question> questionList, boolean isDone) {
         this.name = name;
         this.chapter_id = chapter_id;
         this.questionList = questionList;
         this.isDone = isDone;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -32,7 +32,7 @@ public class Test {
         return name;
     }
 
-    public int getChapter_id() {
+    public long getChapter_id() {
         return chapter_id;
     }
